compile fixes. mostly
- base class for degree config
- edge callback prototype transition
- one more compiler warning
